requirements.txt
Flask
azure-storage-blob
azure.cognitiveservices.vision.computervision
Pillow